renamed Base class to RADDCore and moved to new file (RADD.py). Also, non public __make_dataframes__ and __get_header__ and public get_wts functions are now in RADDCore parent class
